refactor : 카테고리 및 제목 키워드 기준 search 기능 수정 - 작동 테스트 후 기능에 맞게 동작하도록 일부 내용 수정
